Implement basic project.find_up
This version simply looks for the `myproject` directory with a file
named `main` in it, and populates the passed in array reference with the
project root directory and the myproject root directory associated with
it.

In setting up the tests, I copy the test data into a temporary
directory. This prevents `project.find_up` from finding the `myproject`
directory for the MyProject itself.