Improve handling of MYPROJECT_MAPPING env variable
Factored out common functions for testing if a directory is a valid task
definition directory. Created a function to load the mapping
configuration. Created a function to add a mapping to the configuration
if the project directory exists and the myproject directory exists and
is valid.

Added tests to ensure invalid paths fail to be added to the mapping.

Refactored the tests to use some constants for test data paths to make
their use a bit easier.