Update coverage reporting
Allow for skipping lines from coverage; mainly for sanity check
statements that should never happen. These are annotated with 'S' on
coverage reports and are indicated with comments like the following,
where the number indicates the number of following lines to skip:

   # coverage skip: 2

Additionally, I now handle multi-line statements properly, and
propagate the coverage count to all of the lines.